残りHPのパーセント表示が正しく出来ていない不具合を解消

+++ Assets/Scripts/Character/EnemySystem/Core/EnemyBrain.cs
@@ -67,7 +67,7 @@ public class EnemyBrain : CharacterBase, IMatchTarget
         else //ボス
         {
             UIManager.Instance.UpdateBossHP(GetCurrentHP()); //スライダー更新
-            UIManager.Instance.UpdateRemainingHP(Mathf.RoundToInt(GetCurrentHP() / _health.MaxHP * 100)); //パーセント表記更新
+            UIManager.Instance.UpdateRemainingHP(Mathf.RoundToInt((float)GetCurrentHP() / _health.MaxHP * 100)); //パーセント表記更新
         }
         
         Animator.SetTrigger("Damage");
